. Splat map . Fix du bug qui empechait de supprimer un objet en mode map edition . Fix du bug qui empechait le spawn d'avoir le meme comportement en mode map qu'en mode world . Fix du bug d'aplanissement sous un petit objet sur un grand terrain (probleme de précision) . On peut maintenant positionner précisément un objet dans l'editeur de map avant d'aplanir le terrain . Fix du bug des terrains plats visuellement (probleme de normales) .

+++ EasyEngineV4/EasyEngineSrc/Collision/HeightMap.h
@@ -36,7 +36,7 @@ private:
 	float	GetHeight(const CVector& p);
 	float 	SampleHeightmapBilinear(float xModel, float zModel);
 	void	ModelToMap(float xModel, float zModel, float& xMap, float& yMap);
-	void	MapToModel(int xMap, int yMap, float& xModel, float& zModel);
+	void	MapToModel(float xMap, float yMap, float& xModel, float& zModel);
 	void	SetPixelValue(int x, int y, float value);
 	float	GetHeightFromPixelValue(float pixelValue);
 	float	GetPixelValueFromHeight(float height);
